+++ src/engine/engine.h
@@ -9,21 +9,6 @@
 class CFutureLogger;
 class ILogger;
 
-class CHostLookup : public IJob
-{
-private:
-	void Run() override;
-
-public:
-	CHostLookup();
-	CHostLookup(const char *pHostname, int Nettype);
-
-	int m_Result;
-	char m_aHostname[128];
-	int m_Nettype;
-	NETADDR m_Addr;
-};
-
 class IEngine : public IInterface
 {
 	MACRO_INTERFACE("engine", 0)
