+++ src/engine/shared/engine.cpp
@@ -10,19 +10,6 @@
 #include <engine/shared/network.h>
 #include <engine/storage.h>
 
-CHostLookup::CHostLookup() = default;
-
-CHostLookup::CHostLookup(const char *pHostname, int Nettype)
-{
-	str_copy(m_aHostname, pHostname);
-	m_Nettype = Nettype;
-}
-
-void CHostLookup::Run()
-{
-	m_Result = net_host_lookup(m_aHostname, &m_Addr, m_Nettype);
-}
-
 class CEngine : public IEngine
 {
 public:
