+++ src/engine/shared/host_lookup.cpp
@@ -1,0 +1,19 @@
+/* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
+/* If you are missing that file, acquire a complete release at teeworlds.com.                */
+
+#include "host_lookup.h"
+
+#include <base/system.h>
+
+CHostLookup::CHostLookup() = default;
+
+CHostLookup::CHostLookup(const char *pHostname, int Nettype)
+{
+	str_copy(m_aHostname, pHostname);
+	m_Nettype = Nettype;
+}
+
+void CHostLookup::Run()
+{
+	m_Result = net_host_lookup(m_aHostname, &m_Addr, m_Nettype);
+}
